move crawling to DBmanageapp, add dbsqlite code in settings.py
