+++ DBmanageapp/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Dart, Company, Year, Quarter, FS_LoB, FS_Div, FS_Account, SUB_Account
+from .models import *
 # Register your models here.
 admin.site.register(Dart)
 admin.site.register(Company)
