+++ DBmanageapp/models.py
@@ -2,11 +2,11 @@ from django.db import models
 
 
 class Daily_Price(models.Model):
-    open = models.IntegerField(help_text="", null=True, blank=True)
-    high = models.IntegerField(help_text="", null=True, blank=True)
-    low = models.IntegerField(help_text="", null=True, blank=True)
-    close = models.IntegerField(help_text="", null=True, blank=True)
-    volume = models.IntegerField(help_text="", null=True, blank=True)
+    open = models.IntegerField(help_text="시가", null=True, blank=True)
+    high = models.IntegerField(help_text="고가", null=True, blank=True)
+    low = models.IntegerField(help_text="저가", null=True, blank=True)
+    close = models.IntegerField(help_text="종가", null=True, blank=True)
+    volume = models.IntegerField(help_text="거래량", null=True, blank=True)
 
 
 class Kospi(models.Model):
@@ -17,8 +17,8 @@ class Kosdaq(models.Model):
 
 class Company(models.Model):
     price = models.ForeignKey(Daily_Price, on_delete=models.CASCADE, null=True, blank=True)
-    kospi = models.ForeignKey(Kospi, blank=True, null=True, on_delete=models.CASCADE)
-    kosdaq = models.ForeignKey(Kosdaq, blank=True, null=True, on_delete=models.CASCADE)
+    kospi = models.ForeignKey(Kospi, blank=True, null=True)
+    kosdaq = models.ForeignKey(Kosdaq, blank=True, null=True)
     company_name = models.CharField(max_length=200, null=False)
     short_code = models.CharField(max_length=200, blank=True, null=True)
